requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
opencv-python-headless>=4.7.0
numpy>=1.24.0
pandas>=1.5.0
openpyxl>=3.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
tensorboard>=2.13.0
